Further towards memories
Initial approach was to collect $memwr and $memrd cells.  This worked
to some extent and we could build 1rd/1wr port memories.

Playing more with yosys, it seems to make more sense to work with
$mem cells after a 'memory -nobram' pass.  I can probably also
support $memwr/$memrd but I am not sure it's worth it as they
will still require a memory pass ('memory -dff') to work.

The remaining issue is tricky - hardcaml supports 1rd/1wr port
memories natively ONLY.  This is a bit of a limitation but is
what you can easily and robustly describe in RTL.  For yosys we
really need to support multi-port memories.  Further, we need
read-before-write and write-before-read behaviour properly
specified per-port and also write conflict resolution.

We can do all this with LVT multi-port memories, though the
current implemention is a bit rough and ready and will need
to be updated properly (fix RBW behaviour, add per port fallthrough,
add write conflict resolution).

Note; this is just so things work in the general case - in most
situations these things will resolve down to a simple dual port
or single port ram.