A few more features for the lvt based n-port mems
* Support synchronous and asynchronous read ports
  in either read-before-write or write-before-read
  mode (4 modes in total)

* Each read port may have a different read mode
  independantly of the others

* Seperate read/write clocks

Note that in general read-before-write and
write-before-read only really seems to make sense
when the read and write ports are in the same clock
domain.  Not sure what happens with real hardware
implementations of these sorts of things but I cant
get my head round any sort of behaviour you could
rely on in this case.

As such the various read/write/async/sync behaviours
are only specified when the read and write clocks
are in the same domain.  Some restrictions are
lifted when there is only one write port and the
memory is synchronous (this sort of mapping is most
useful as it matches the general behaviour of fpga
memories).

Finally, although a write clock is specified per
write port, only the 1st one is actually used (the
LVT ram which coordinates accesses is a bank of
registers and only 1 clock makes sense).

Finally, integration.  We are still missing the
per-bit write enable feature - we can build one
on top (though we will have to replicate the LVT
unnecessarily so we should really do it internally).
Also, this should really replace the buggy version
in HardCaml itself at some point.  Also, when some
of the assumptions list above are not valid, we
should assert or warn.